pyproject.toml
[tool.poetry]
name = "expense-precheck"
version = "0.1.0"
description = "AI-assisted expense approval precheck system with document-aware validation."
authors = ["Finance Automation Team <finance-automation@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pydantic = "^2.8.0"
streamlit = "^1.38.0"
pandas = "^2.2.0"
requests = "^2.32.0"
httpx = "^0.27.0"
PyMuPDF = "^1.24.0"
python-docx = "^1.1.0"
PyYAML = "^6.0.0"
python-dateutil = "^2.9.0"
Pillow = "^10.2.0"
pytesseract = "^0.3.10"
beautifulsoup4 = "^4.12.0"

[tool.poetry.group.dev.dependencies]
black = "^24.8"
ruff = "^0.6.5"
pytest = "^8.3.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
